New class: CoroAsyncHelper, to help with waiting on asnyc coroutine results
The coroutine can give an instance of this class to an external object, and it
can the "wait" on the asynchronous result. When the external object finished
processing, it can "wake up" the coroutine, which can then process the results.
The results of the asynchronous processing should be stored in the classes
derived from CoroAsyncHelper.